requirements.txt
httpx>=0.23.0
fzf.py @ git+https://github.com/justfoolingaround/fzf.py
regex>=2022.3.15
